added the animations of the doors opening
added the animations of the doors opening. was orginally trying to do it to the door prefab but it wasn't working with the position because it was using the prefabs the position and because the doors have different positions it was making the door in scene would spam to the position that the prefab in the prefab window was and then play the animation. so i instead animated each door